Try to initialize known_hosts.
There are at least three ip addresses, so calling twice isn't
sufficient. However if we have at least one it seems to add
the other addresses without prompting. If we don't have an
entry it asks for confirmation, so try connecting with the
check disabled when we build the container.